[FEAT] add floor model for rooms
This commit adds the Floor model, factory, and Livewire component. The Floor model represents a floor in a building and has the attributes 'denomination' and 'description'. The Floor factory is responsible for generating dummy data for the Floor model. The Livewire component 'Floor' is used to display and search for floors in the admin room section. This commit also includes the migration file to create the 'floors' table in the database and the seeder to populate it with initial data.